+++ changelog.md
@@ -49,3 +49,4 @@
 - update creating Delivery type of request.
 - add field `category` and `liked` to type `Outfit`. Create and input `OutfitInput` and created mutation `updateUserOutfitMutation`, `addItemToOutfit`, `likeAnOutfit` and `unlikeAnOutfit`.
 - Fix big on `makePayment` to return current subscription plan of the user.
+- 🐛 FIX: add tag field to itemInput and likeOutfit, unlikeOutfit return Outfit object and sendOutRequest return Request object